add assertion for cast double->float
ScoreFunction scoring might result in under or overflow, for example if a user
decides to use the timestamp as a boost in the script scorer. Therefore, check
if cast causes a huge precision loss. Note that this does not always detect
casting issues. For example in
ScriptFunction.score()
the function
SearchScript.runAsDouble()
is called. AbstractFloatSearchScript implements it as follows:
@Override     public double runAsDouble() {         return runAsFloat();     }
In this case the cast happens before the assertion and therfore precision
lossor over/underflows cannot be detected by the assertion.

+++ src/main/java/org/elasticsearch/common/lucene/search/function/BoostScoreFunction.java
@@ -33,7 +33,6 @@ public class BoostScoreFunction implements ScoreFunction {
         this.boost = boost;
     }
 
-
     public float getBoost() {
         return boost;
     }
@@ -44,7 +43,7 @@ public class BoostScoreFunction implements ScoreFunction {
     }
 
     @Override
-    public float score(int docId, float subQueryScore) {
+    public double score(int docId, float subQueryScore) {
         return subQueryScore * boost;
     }
 
@@ -68,12 +67,15 @@ public class BoostScoreFunction implements ScoreFunction {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
 
         BoostScoreFunction that = (BoostScoreFunction) o;
 
-        if (Float.compare(that.boost, boost) != 0) return false;
+        if (Float.compare(that.boost, boost) != 0)
+            return false;
 
         return true;
     }
